# internal/gen/api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Judex API
  version: 1.0.0
  summary: Judex 协作平台公共 API（docs/plans/v1/06 契约源）
  description: |
    所有业务路径前缀 /api/v1（本规范路径不含该前缀，由网关/服务器统一挂载）。

    协议规则：JSON 字段 camelCase；UUID 字符串；时间 RFC3339 UTC；版本为正整数。
    成功 envelope `{data, meta:{requestId,serverTime,eventCursor?}}`；命令额外返回 `affected`。
    错误 envelope `{error:{code,message,details?,retryable},requestId}`；客户端按 code+details 处理，不解析 message。
    分页 `?limit=50&cursor=<opaque>`（1-100）；列表 data 为 `{items,nextCursor}`。
    有副作用的业务命令要求 Idempotency-Key 头（UUID）；已有对象写操作要求 expectedVersion；正式决定要求 reviewId/reviewHash。
    Web 会话使用 Cookie+CSRF（sessionCookie）；CLI 使用受限 Bearer grant（bearerGrant）。
    阶段实现状态由 GET /system capabilities 返回；未实现的 operation 返回 501 而不是假成功。
servers:
  - url: /api/v1
    description: Same-origin (SPA or CLI with explicit base URL)
...
security:
  - sessionCookie: []
  - bearerGrant: []
paths:
...
  /projects/{projectId}/material-versions/{versionId}:
    get:
      operationId: getMaterialVersionById
      summary: 按固定版本读取材料清单
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: versionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 不可变材料版本与清单
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/MaterialVersion'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/material-versions/{versionId}/content:
    get:
      operationId: downloadMaterialVersionById
      summary: 版本内容下载（鉴权代理，支持 Range）
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: versionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Range
          in: header
          required: false
          schema:
            type: string
        - name: entry
          in: query
          required: false
          description: HTML 包内的相对路径条目。
          schema:
            type: string
      responses:
        '200':
          description: 内容流。
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: 范围内容。
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /me/sessions:
    get:
      operationId: listSessions
      summary: 本人活跃会话列表（无 secret）
      tags:
        - me
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: 会话分页。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - items
                      - nextCursor
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/SessionSummary'
                      nextCursor:
                        type:
                          - string
                          - 'null'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /projects/{projectId}/uploads:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      operationId: listUploadSessions
      summary: 本人未完成上传会话
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      responses:
        '200':
          description: 会话分页。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - items
                      - nextCursor
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/UploadSession'
                      nextCursor:
                        type:
                          - string
                          - 'null'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createUploadSession
      summary: 创建上传会话（返回分片方案）
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadRequest'
      responses:
        '201':
          description: 上传方案。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/UploadSession'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/TooLarge'
  /projects/{projectId}/uploads/{uploadId}/parts/{partNumber}:
    put:
      operationId: uploadPart
      summary: 上传单个分片（二进制流；同号不同摘要冲突）
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: partNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: X-Judex-Part-SHA256
          in: header
          required: true
          schema:
            type: string
            pattern: ^[0-9a-f]{64}$
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: 分片已登记。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - partNumber
                      - receivedBytes
                    properties:
                      partNumber:
                        type: integer
                      receivedBytes:
                        type: integer
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/TooLarge'
  /projects/{projectId}/uploads/{uploadId}/complete:
    post:
      operationId: completeUpload
      summary: 完成上传并登记不可变材料版本（幂等返回同版本）
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteUploadRequest'
      responses:
        '201':
          description: 已登记的材料版本。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/MaterialVersion'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 来源版本落后（SOURCE_VERSION_CONFLICT）或幂等冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '412':
          description: 分片不完整或校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{projectId}/uploads/{uploadId}:
    delete:
      operationId: cancelUpload
      summary: 取消未完成上传会话（不删正式材料）
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 已取消。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - cancelled
                    properties:
                      cancelled:
                        type: boolean
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/materials:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      operationId: listMaterials
      summary: 项目共享资料分页（按 kind/关联过滤）
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: kind
          in: query
          schema:
            type: string
            enum:
              - file
              - html_bundle
              - text
        - name: linkedObjectType
          in: query
          schema:
            type: string
        - name: linkedObjectId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 材料分页。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - items
                      - nextCursor
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Material'
                      nextCursor:
                        type:
                          - string
                          - 'null'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}/materials/{materialId}/versions:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: materialId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: listMaterialVersions
      summary: 不可变版本列表
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      responses:
        '200':
          description: 版本分页。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - items
                      - nextCursor
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/MaterialVersion'
                      nextCursor:
                        type:
                          - string
                          - 'null'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/materials/{materialId}/versions/{versionId}/content:
    get:
      operationId: downloadMaterialContent
      summary: 版本内容下载（鉴权代理，支持 Range）
      tags:
        - materials
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: materialId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: versionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Range
          in: header
          required: false
          schema:
            type: string
        - name: entry
          in: query
          required: false
          description: HTML 包内的相对路径条目。
          schema:
            type: string
      responses:
        '200':
          description: 内容流。
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: 范围内容。
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/materials/{materialId}/versions/{versionId}/preview-session:
    post:
      operationId: createPreviewSession
      summary: 为 HTML 包版本创建隔离 origin 短时预览
      tags:
        - materials
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: materialId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: versionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      responses:
        '201':
          description: 预览会话。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/PreviewSession'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  schemas:
...
    SessionSummary:
      type: object
      required:
        - id
        - createdAt
        - expiresAt
        - lastSeenAt
        - current
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
        current:
          type: boolean
...
    RequestMeta:
      type: object
      required:
        - requestId
        - serverTime
      properties:
        requestId:
          type: string
        serverTime:
          type: string
          format: date-time
        eventCursor:
          type: integer
          format: int64
          description: 当前项目事件序列游标，仅项目资源响应返回。
    AffectedResource:
      type: object
      required:
        - type
        - id
        - version
      properties:
        type:
          type: string
        id:
          type: string
        version:
          type: integer
          minimum: 1
    CommandMeta:
      allOf:
        - $ref: '#/components/schemas/RequestMeta'
        - type: object
          properties:
            affected:
              type: array
              items:
                $ref: '#/components/schemas/AffectedResource'
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 兜底文案；客户端按 code+details 映射，不解析句子。
        details:
          description: 结构化错误详情，按 code 不同而不同。
        retryable:
          type: boolean
          default: false
...
    ErrorResponse:
      type: object
      required:
        - error
        - requestId
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        requestId:
          type: string
...
    UploadSession:
      type: object
      required:
        - id
        - state
        - kind
        - expectedSize
        - checksum
        - partSize
        - expiresAt
      properties:
        id:
          type: string
          format: uuid
        state:
          type: string
          enum:
            - open
            - completed
            - cancelled
            - expired
        kind:
          type: string
          enum:
            - file
            - html_bundle
        name:
          type: string
        expectedSize:
          type: integer
        checksum:
          type: string
          description: 完整文件 SHA256 十六进制。
        mime:
          type: string
        partSize:
          type: integer
          description: 分片大小字节数（推荐 8MiB）。
        partCount:
          type: integer
        materialId:
          type:
            - string
            - 'null'
          format: uuid
        resultVersionId:
          type:
            - string
            - 'null'
          format: uuid
        expiresAt:
          type: string
          format: date-time
    CreateUploadRequest:
      type: object
      required:
        - name
        - size
        - sha256
        - mime
        - kind
      properties:
        name:
          type: string
          maxLength: 500
        size:
          type: integer
          minimum: 1
        sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
        mime:
          type: string
        kind:
          type: string
          enum:
            - file
            - html_bundle
        entrypoint:
          type: string
          description: HTML 包入口相对路径。
    CompleteUploadRequest:
      type: object
      required:
        - parts
      properties:
        parts:
          type: array
          items:
            type: object
            required:
              - partNumber
              - sha256
            properties:
              partNumber:
                type: integer
                minimum: 1
              sha256:
                type: string
        sourceVersion:
          type: integer
          description: 针对已有材料派生发布时的当前版本号；落后返回 SOURCE_VERSION_CONFLICT。
    Material:
      type: object
      required:
        - id
        - title
        - kind
        - visibility
        - currentVersionId
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        kind:
          type: string
          enum:
            - file
            - html_bundle
            - text
        visibility:
          type: string
          enum:
            - project
        currentVersionId:
          type:
            - string
            - 'null'
          format: uuid
        createdAt:
          type: string
          format: date-time
    MaterialVersion:
      type: object
      required:
        - id
        - revision
        - state
        - sha256
        - size
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        materialId:
          type: string
          format: uuid
        revision:
          type: integer
          minimum: 1
        state:
          type: string
          enum:
            - staged
            - ready
            - unavailable
        sha256:
          type: string
        size:
          type: integer
        mime:
          type: string
        entrypoint:
          type:
            - string
            - 'null'
        entries:
          type: array
          description: HTML 包的相对路径清单。
          items:
            type: object
            required:
              - relativePath
              - size
              - sha256
            properties:
              relativePath:
                type: string
              size:
                type: integer
              sha256:
                type: string
        authorId:
          type:
            - string
            - 'null'
        runId:
          type:
            - string
            - 'null'
        createdAt:
          type: string
          format: date-time
    PreviewSession:
      type: object
      required:
        - previewUrl
        - expiresAt
      properties:
        previewUrl:
          type: string
          format: uri
          description: 隔离 preview origin 的短时 URL。
        expiresAt:
          type: string
          format: date-time
...
  parameters:
    LimitParam:
      name: limit
      in: query
      description: 分页大小，范围 1-100，默认 50。
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    CursorParam:
      name: cursor
      in: query
      description: 不透明分页游标，编码范围与过滤条件。
      schema:
        type: string
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      required: true
      description: 客户端生成的 UUID；同键同载荷重试返回原结果，同键异载荷返回 409。
      schema:
        type: string
        format: uuid
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: 未认证（401，UNAUTHENTICATED/SESSION_EXPIRED/GRANT_REVOKED）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 已认证但无权限（403，FORBIDDEN/HUMAN_CONFIRMATION_REQUIRED）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在或对非成员隐藏存在性（404）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 状态/版本/幂等冲突（409，详见 error.code）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooLarge:
      description: 请求体过大或类型不支持（413/415）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-judex_session
      description: Web 会话 Cookie（本地 HTTP 开发使用 judex_dev_session）。
    bearerGrant:
      type: http
      scheme: bearer
      description: CLI access token（受限 grant，无人工审批能力）。